requirements.txt
bs4
selenium
requests
sqlalchemy
python-dotenv
pymysql
cryptography